+++ chess/__init__.py
@@ -24,7 +24,7 @@ __author__ = "Niklas Fiekas"
 
 __email__ = "niklas.fiekas@backscattering.de"
 
-__version__ = "0.30.1"
+__version__ = "0.31.0"
 
 import collections
 import copy
@@ -2925,11 +2925,11 @@ class Board(BaseBoard):
         Checks if the given pseudo-legal move is irreversible.
 
         In standard chess, pawn moves, captures, moves that destroy castling
-        rights and moves the cede en-passant are irreversible.
+        rights and moves that cede en passant are irreversible.
 
-        This method has false-negatives with forced lines: For example a check
+        This method has false-negatives with forced lines. For example, a check
         that will force the king to lose castling rights is not considered
-        irreversible. The actual king move is.
+        irreversible. Only the actual king move is.
         """
         return self.is_zeroing(move) or self._reduces_castling_rights(move) or self.has_legal_en_passant()