+++ CHANGELOG.rst
@@ -17,6 +17,9 @@ Changes:
 
 Bugfixes:
 
+* `chess.Board.is_irreversible()` now considers ceding legal en-passant
+  captures as irreversible. Also documented that false-negatives due to forced
+  lines are by design.
 * Fixed stack overflow in `chess.pgn` when exporting, visiting or getting the
    final board of a very long game.
 * Clarified documentation regarding board validity.